#7.4 Protecting Video Routes

+++ controllers/videoController.js
@@ -43,9 +43,11 @@ export const postUpload = async (req, res) => {
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
-  console.log(newVideo);
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -54,7 +56,8 @@ export const videoDetail = async (req, res) => {
     params: { id }
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator");
+    console.log(video);
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     console.log(error);
@@ -68,6 +71,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
+    // console.log("1>>>" + typeof req.user.id);
+    // console.log("2>>>" + typeof video.creator);
+    if (video.creator.toString() !== req.user.id) {
+      throw Error();
+    }
     res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
   } catch (error) {
     console.log(error);
@@ -95,6 +103,10 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
+    const video = await Video.findById(id);
+    if (video.creator.toString() !== req.user.id) {
+      throw Error();
+    }
     await Video.findOneAndRemove({ _id: id });
   } catch (error) {
     console.log(error);
